+++ poet/locations.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from pip.utils.appdirs import user_cache_dir
+from pip.utils.appdirs import user_cache_dir, user_config_dir
 
 
 CACHE_DIR = user_cache_dir('pypoet')
+CONFIG_DIR = user_config_dir('pypoet')